% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClumppOpts.R
\docType{class}
\name{ClumppOpts-class}
\alias{ClumppOpts-class}
\title{ClumppOpts: An S4 class to represent Structure parameters}
\description{
This class stores input parameters for the Structure program.
}
\section{Slots}{

\describe{
\item{\code{M}}{\code{character} name of search method. Valid arguments are 'FullSearch', 'Greedy', or 'LargeKGreedy'. Defaults to 'Greedy'.}

\item{\code{W}}{\code{logical} weight populations by number of individuals? Defaults to \code{TRUE}.}

\item{\code{S}}{\code{logical} if \code{TRUE} the \code{G} matrix similarity statistic ised used. Else the \code{G\\'} statistic is used. Defaults to code{FALSE}.}

\item{\code{REPEATS}}{\code{numeric} Number of random input orders tested. Defaults to 1000.}
}}
\seealso{
\code{\link{ClumppOpts}}.
}

