% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureCollection.R
\name{delta.k.plot}
\alias{delta.k.plot}
\alias{delta.k.plot.StructureCollection}
\title{Plot delta-K values}
\usage{
delta.k.plot(x, main)

\method{delta.k.plot}{StructureCollection}(x, main = "Delta-K plot")
}
\arguments{
\item{x}{\code{StructureCollection} object.}

\item{main}{\code{character} plot title.}
}
\description{
This function plots the delta-K values in a \code{StructureCollection} object.
}
\seealso{
\code{\link{StructureCollection}}.
}

