% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{gelman.diag}
\alias{gelman.diag}
\alias{gelman.diag.StructureAnalysis}
\alias{gelman.diag.StructureCollection}
\alias{gelman.diag.default}
\title{Gelman-Rubin diagnostic statistics}
\usage{
gelman.diag(x, ...)

\method{gelman.diag}{default}(x, ...)

\method{gelman.diag}{StructureAnalysis}(x, ...)

\method{gelman.diag}{StructureCollection}(x, K = x@best, ...)
}
\arguments{
\item{x}{\code{StructureCollection} or \code{StructureAnalysis} object.}

\item{...}{arguments passed to \code{\link[code]{gelman.diag}}.}
}
\description{
This function returns the Gelman-Rubin diagnostic statistics for the negative loglikelihood of multiple Structure runs. See \code{\link[code]{gelman.dig}} for more information.
}

