% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureCollection.R
\name{lnprob.plot}
\alias{lnprob.plot}
\alias{lnprob.plot.StructureCollection}
\title{Plot estimated Ln probability of data}
\usage{
lnprob.plot(x, main, ...)

\method{lnprob.plot}{StructureCollection}(x, main = "", ...)
}
\arguments{
\item{x}{\code{StructureCollection} object.}

\item{main}{\code{character} plot title.}

\item{...}{not used.}
}
\description{
This function plots the estimated Ln probability of the data assuming different number of populations in a \code{StructureCollection} object.
}
\seealso{
\code{\link{StructureCollection}}.
}

