% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureAnalysis.R
\name{lnprob}
\alias{lnprob}
\alias{lnprob.StructureAnalysis}
\alias{lnprob.StructureReplicate}
\alias{lnprob.StructureResults}
\title{Probability of the model}
\usage{
lnprob(x)

\method{lnprob}{StructureReplicate}(x)

\method{lnprob}{StructureResults}(x)

\method{lnprob}{StructureAnalysis}(x)
}
\arguments{
\item{x}{\code{StructureReplicate}, \code{StructureResults}, \code{StructureAnalysis} or \code{StructureCollection} object.}
}
\value{
\code{numeric}.
}
\description{
This function returns the ln probability of a Structure analysis.
}

