% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureData.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{loci.names}
\alias{loci.names}
\alias{loci.names.StructureAnalysis}
\alias{loci.names.StructureCollection}
\alias{loci.names.StructureData}
\alias{loci.names<-}
\alias{loci.names<-.StructureAnalysis}
\alias{loci.names<-.StructureCollection}
\alias{loci.names<-.StructureData}
\title{Loci names}
\usage{
loci.names(x)

loci.names(x) <- value

\method{loci.names}{StructureData}(x)

\method{loci.names}{StructureData}(x) <- value

\method{loci.names}{StructureAnalysis}(x)

\method{loci.names}{StructureAnalysis}(x) <- value

\method{loci.names}{StructureCollection}(x)

\method{loci.names}{StructureCollection}(x) <- value
}
\arguments{
\item{x}{\code{StructureData}, \code{StructureAnalysis}.}

\item{value}{\code{character} new loci names.}
}
\value{
\code{character}.
}
\description{
This function returns the loci names in a \code{Structure} object.
}

