% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureData.R
\name{loci.subset}
\alias{loci.subset}
\alias{loci.subset.StructureData}
\title{Subset loci}
\usage{
loci.subset(x, loci)

\method{loci.subset}{StructureData}(x, loci)
}
\arguments{
\item{x}{\code{StructureData}, or \code{StructureAnalysis}.}

\item{loci}{\code{character}, \code{numeric}, or code{logical} indicating the loci to return.}
}
\value{
\code{StructureData}.
}
\description{
This function returns a subset of loci in a \code{Structure} object.
}

