% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureCollection.R
\name{loglik.plot}
\alias{loglik.plot}
\alias{loglik.plot.StructureCollection}
\title{Plot negative log-likelihood values}
\usage{
loglik.plot(x, main)

\method{loglik.plot}{StructureCollection}(x, main = "")
}
\arguments{
\item{x}{\code{StructureCollection} object.}

\item{main}{\code{character} plot title.}
}
\description{
This function plots the negative log-likelihood values in a \code{StructureCollection} object.
}
\seealso{
\code{\link{StructureCollection}}.
}

