% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureAnalysis.R
\name{logLik}
\alias{logLik}
\alias{logLik.StructureAnalysis}
\alias{logLik.StructureReplicate}
\alias{logLik.StructureResults}
\title{Negative Log Likelihood}
\usage{
\method{logLik}{StructureReplicate}(object, ...)

\method{logLik}{StructureResults}(object, ...)

\method{logLik}{StructureAnalysis}(object, ...)
}
\arguments{
\item{object}{\code{StructureReplicate}, \code{StructureResults}, \code{StructureAnalysis} or \code{StructureCollection} object.}

\item{...}{not used.}
}
\value{
\code{numeric}.
}
\description{
This function returns the negative loglikelihood of a Structure analysis.
}

