% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureData.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{n.loci}
\alias{n.loci}
\alias{n.loci.StructureAnalysis}
\alias{n.loci.StructureCollection}
\alias{n.loci.StructureData}
\title{Number of loci}
\usage{
n.loci(x)

\method{n.loci}{StructureData}(x)

\method{n.loci}{StructureAnalysis}(x)

\method{n.loci}{StructureCollection}(x)
}
\arguments{
\item{x}{\code{StructureData}, \code{StructureAnalysis}.}
}
\value{
\code{integer}.
}
\description{
This function returns the number of loci in a \code{Structure} object.
}

