% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureOpts.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{n.pop}
\alias{n.pop}
\alias{n.pop.StructureAnalysis}
\alias{n.pop.StructureCollection}
\alias{n.pop.StructureOpts}
\alias{n.pop.StructureReplicate}
\alias{n.pop.StructureResults}
\title{Number of populations}
\usage{
n.pop(x)

\method{n.pop}{StructureReplicate}(x)

\method{n.pop}{StructureResults}(x)

\method{n.pop}{StructureOpts}(x)

\method{n.pop}{StructureAnalysis}(x)

\method{n.pop}{StructureCollection}(x)
}
\arguments{
\item{x}{\code{StructureReplicate}, \code{StructureResults}, or \code{StructureAnalysis}.}
}
\value{
\code{integer}.
}
\description{
This function returns the number of populations sin a \code{Structure} object.
}

