% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureData.R, R/StructureOpts.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{print}
\alias{print}
\alias{print.StructureAnalysis}
\alias{print.StructureCollection}
\alias{print.StructureData}
\alias{print.StructureOpts}
\alias{print.StructureReplicate}
\alias{print.StructureResults}
\title{Print objects}
\usage{
\method{print}{StructureReplicate}(x, ..., header = TRUE)

\method{print}{StructureResults}(x, ..., header = TRUE)

\method{print}{StructureData}(x, ..., header = TRUE)

\method{print}{StructureOpts}(x, ..., header = TRUE)

\method{print}{StructureAnalysis}(x, ..., header = TRUE)

\method{print}{StructureCollection}(x, ..., header = TRUE)
}
\arguments{
\item{x}{the object to print.}

\item{...}{not used.}

\item{header}{\code{logical} should object header be shown?}
}
\description{
This function prints objects.
}

