% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurer-internal.R
\name{read.ClumppReplicates}
\alias{read.ClumppReplicates}
\title{Read replicate data to file for clumpp}
\usage{
read.ClumppReplicates(file)
}
\arguments{
\item{file}{\code{character} file path to save data.}
}
\value{
\code{list} of \code{StructureReplicate} objects.
}
\description{
This function reads results from clumpp.
}

