% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureData.R
\name{read.StructureData}
\alias{read.StructureData}
\title{Read FSTAT data for Structure}
\usage{
read.StructureData(x)
}
\arguments{
\item{x}{\code{character} file path with data}
}
\value{
\code{StructureData}.
}
\description{
This function reads FSTAT data for Structure.
}
\examples{
x <- read.StructureData(system.file('extdata', 'example_fstat_aflp.dat', package='structurer'))
}
\seealso{
\code{StructureData}.
}

