% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureReplicate.R
\name{read.StructureReplicate}
\alias{read.StructureReplicate}
\title{Read Structure run}
\usage{
read.StructureReplicate(file, runfile)
}
\arguments{
\item{file}{\code{character} file path of output file.}

\item{runfile}{\code{character} file path of file text printed to console during run.}
}
\value{
\code{\link{StructureReplicate}}.
}
\description{
This function reads the results of a single run of the Structure program.
}
\seealso{
\code{\link{StructureReplicate-class}}.
}

