% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureAnalysis.R
\name{run.single.Structure}
\alias{run.single.Structure}
\title{Run Structure anaylsis for a single MAXPOPS parameter.}
\usage{
run.single.Structure(x, NUMRUNS = 2, MAXPOPS = 2, BURNIN = 10000,
  NUMREPS = 20000, NOADMIX = FALSE, ADMBURNIN = 500, SEED = NA_real_,
  M = "Greedy", W = TRUE, S = FALSE, REPEATS = 1000, dir = tempdir(),
  clean = TRUE, verbose = FALSE, threads = 1)
}
\arguments{
\item{x}{\code{StructureData} object.}

\item{NUMRUNS}{\code{numeric} Number of replicate Structure runs. Defaults to 2.}

\item{MAXPOPS}{\code{numeric} Number of populations assumed. Defaults to 2.}

\item{BURNIN}{\code{numeric} Length of burnin period. Defaults to 10000.}

\item{NUMREPS}{\code{numeric} Number of MCMC iterations for inference. Defaults to 20000.}

\item{NOADMIX}{\code{logical} Do not use admixture model. Defaults to \code{FALSE}.}

\item{ADMBURNIN}{\code{numeric} Initial period of burnin with admixture model. Defaults to 500.}

\item{SEED}{\code{numeric} Seed for random number generator. Defaults to NA so a random seed is used.}

\item{M}{\code{character} name of search method. Valid arguments are 'FullSearch', 'Greedy', or 'LargeKGreedy'. Defaults to 'Greedy'.}

\item{W}{\code{logical} weight populations by number of individuals? Defaults to \code{TRUE}.}

\item{S}{\code{logical} if \code{TRUE} the \eqn{G} matrix similarity statistic ised used. Else the \eqn{G \prime} statistic is used. Defaults to code{FALSE}.}

\item{REPEATS}{\code{numeric} Number of random input orders tested. Defaults to 1000.}

\item{dir}{\code{character} with directory to use for analysis.}

\item{clean}{\code{logical} should input and output files be deleted after analysis is finished?}

\item{verbose}{\code{logical} should messages be printed during processing?}

\item{threads}{\code{numeric} number of threads to use for processing. Defaults to 1.}
}
\description{
This function analyses data using Structure. It generates replicate runs for a single MAXPOPS parameter.
}
\examples{
# run Structure using low number of iterations
dat <- read.StructureData(system.file('extdata', 'example_fstat_aflp.dat', package='structurer'))
x <- run.single.Structure(dat, NUMRUNS=1, MAXPOPS=2, BURNIN=10,
NUMREPS=10, NOADMIX=FALSE, ADMBURNIN=10)
}
\seealso{
\code{StructureData}, \code{StructureOpts}.
}

