% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureCollection.R
\name{run.Structure}
\alias{run.Structure}
\alias{run.Structure.StructureData}
\alias{run.Structure.list}
\title{Run Structure anaylsis}
\usage{
run.Structure(x, NUMRUNS, MAXPOPS, BURNIN, NUMREPS, NOADMIX, ADMBURNIN,
  FREQSCORR, UPDATEFREQ, M, W, S, REPEATS, dir, clean, verbose, threads, SEED)

\method{run.Structure}{StructureData}(x, NUMRUNS = 2, MAXPOPS = 1:10,
  BURNIN = 10000, NUMREPS = 20000, NOADMIX = FALSE, ADMBURNIN = 500,
  FREQSCORR = TRUE, UPDATEFREQ = max(floor(BURNIN + NUMREPS)/1000, 1),
  M = "Greedy", W = TRUE, S = FALSE, REPEATS = 1000, dir = tempdir(),
  clean = TRUE, verbose = FALSE, threads = 1, SEED = sample.int(1e+05,
  NUMRUNS * length(MAXPOPS)))

\method{run.Structure}{list}(x, NUMRUNS = 2, MAXPOPS = 1:10,
  BURNIN = 10000, NUMREPS = 20000, NOADMIX = FALSE, ADMBURNIN = 500,
  FREQSCORR = TRUE, UPDATEFREQ = max(floor(BURNIN + NUMREPS)/1000, 1),
  M = "Greedy", W = TRUE, S = FALSE, REPEATS = 1000,
  dir = file.path(tempdir(), seq_along(x)), clean = TRUE, verbose = FALSE,
  threads = 1, SEED = sample.int(1e+05, NUMRUNS * length(MAXPOPS) *
  length(x)))
}
\arguments{
\item{x}{\code{StructureData} object.}

\item{NUMRUNS}{\code{numeric} Number of replicate Structure runs. Defaults to 2.}

\item{MAXPOPS}{\code{numeric} Number of populations assumed. Defaults to 2.}

\item{BURNIN}{\code{numeric} Length of burnin period. Defaults to 10000.}

\item{NUMREPS}{\code{numeric} Number of MCMC iterations for inference. Defaults to 20000.}

\item{NOADMIX}{\code{logical} Do not use admixture model. Defaults to \code{FALSE}.}

\item{ADMBURNIN}{\code{numeric} Initial period of burnin with admixture model. Defaults to 500.}

\item{FREQSCORR}{\code{logical} Allele frequencies are correlated among populations? Defaults to \code{TRUE}.}

\item{UPDATEFREQ}{\code{numeric} Frequency to store updates to loglikelihood for traceplots. Defaults to yield 1000 frequencies.}

\item{M}{\code{character} name of search method. Valid arguments are 'FullSearch', 'Greedy', or 'LargeKGreedy'. Defaults to 'Greedy'.}

\item{W}{\code{logical} weight populations by number of individuals? Defaults to \code{TRUE}.}

\item{S}{\code{logical} if \code{TRUE} the \eqn{G} matrix similarity statistic ised used. Else the \eqn{G \prime} statistic is used. Defaults to code{FALSE}.}

\item{REPEATS}{\code{numeric} Number of random input orders tested. Defaults to 1000.}

\item{dir}{\code{character} with directory to use for analysis.}

\item{clean}{\code{logical} should input and output files be deleted after analysis is finished?}

\item{verbose}{\code{logical} should messages be printed during processing?}

\item{threads}{\code{numeric} number of threads to use for processing. Defaults to 1.}

\item{SEED}{\code{numeric} Seed for random number generator. Defaults to NA so a random seed is used.}
}
\description{
This function analyses data using Structure.
}
\examples{
# run Structure using low number of iterations
dat <- read.StructureData(system.file('extdata', 'example_fstat_aflp.dat', package='structurer'))
x <- run.Structure(dat, NUMRUNS=2, MAXPOPS=1:3, BURNIN=10, NUMREPS=10, NOADMIX=FALSE, ADMBURNIN=10)
# run Structure for a list of two StructureData objects
x2 <- run.Structure(list(dat, dat), NUMRUNS=2, MAXPOPS=1:3, BURNIN=10, NUMREPS=10, NOADMIX=FALSE, 
 ADMBURNIN=10)
print(x)
}
\seealso{
\code{StructureData}, \code{StructureOpts}.
}

