% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{sample.membership}
\alias{sample.membership}
\alias{sample.membership.StructureAnalysis}
\alias{sample.membership.StructureCollection}
\alias{sample.membership.StructureReplicate}
\alias{sample.membership.StructureResults}
\title{Sample membership}
\usage{
sample.membership(x)

\method{sample.membership}{StructureReplicate}(x)

\method{sample.membership}{StructureResults}(x)

\method{sample.membership}{StructureAnalysis}(x)

\method{sample.membership}{StructureCollection}(x)
}
\arguments{
\item{x}{\code{StructureReplicate}, \code{StructureResults}, \code{StructureAnalysis}.}
}
\value{
\code{integer}.
}
\description{
This function returns the membership of samples in a \code{Structure} object.
}

