% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureData.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{sample.names}
\alias{sample.names}
\alias{sample.names.StructureAnalysis}
\alias{sample.names.StructureCollection}
\alias{sample.names.StructureData}
\alias{sample.names.StructureReplicate}
\alias{sample.names.StructureResults}
\alias{sample.names<-}
\alias{sample.names<-.StructureAnalysis}
\alias{sample.names<-.StructureCollection}
\alias{sample.names<-.StructureData}
\title{Sample names}
\usage{
sample.names(x)

sample.names(x) <- value

\method{sample.names}{StructureReplicate}(x)

\method{sample.names}{StructureResults}(x)

\method{sample.names}{StructureData}(x)

\method{sample.names}{StructureData}(x) <- value

\method{sample.names}{StructureAnalysis}(x)

\method{sample.names}{StructureAnalysis}(x) <- value

\method{sample.names}{StructureCollection}(x)

\method{sample.names}{StructureCollection}(x) <- value
}
\arguments{
\item{x}{\code{StructureData}, \code{StructureAnalysis}.}

\item{value}{\code{character} vector of new names;}
}
\value{
\code{character}.
}
\description{
This function returns the sample names in a \code{Structure} object.
}

