% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureData.R, R/StructureOpts.R, R/StructureAnalysis.R, R/StructureCollection.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,StructureAnalysis-method}
\alias{show,StructureCollection-method}
\alias{show,StructureData-method}
\alias{show,StructureOpts-method}
\alias{show,StructureReplicate-method}
\alias{show,StructureResults-method}
\title{Show objects}
\usage{
\S4method{show}{StructureReplicate}(object)

\S4method{show}{StructureResults}(object)

\S4method{show}{StructureData}(object)

\S4method{show}{StructureOpts}(object)

\S4method{show}{StructureAnalysis}(object)

\S4method{show}{StructureCollection}(object)
}
\arguments{
\item{object}{the object to show.}
}
\description{
This function shows objects.
}

