% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureAnalysis.R
\docType{class}
\name{StructureAnalysis-class}
\alias{StructureAnalysis-class}
\title{StructureAnalysis: An S4 class to represent inputs and outputs from Structure}
\description{
This class stores input data and associated output results from the Structure program.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{StructureOpts} object with parameters used to run Structure .}

\item{\code{data}}{\code{StructureData} object with input data used for analysis.}

\item{\code{results}}{\code{StructureResults} object with results from analysis.}
}}
\seealso{
\code{\link{StructureAnalysis}}.
}

