% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureAnalysis.R
\name{StructureAnalysis}
\alias{StructureAnalysis}
\title{Create StructureAnalysis object}
\usage{
StructureAnalysis(opts, data, results)
}
\arguments{
\item{opts}{\code{StructureOpts} object with parameters used to run Structure .}

\item{data}{\code{StructureData} object with input data used for analysis.}

\item{results}{\code{StructureResults} object with results from analysis.}
}
\description{
This function creates a new \code{StructureAnalysis} object.
}
\seealso{
\code{\link{StructureAnalysis-class}}, \code{\link{StructureData}}, \code{\link{StructureData}}, \code{\link{StructureResults}}.
}

