% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureCollection.R
\docType{class}
\name{StructureCollection-class}
\alias{StructureCollection-class}
\title{StructureCollection: An S4 class to represent a collection of inputs and outputs from Structure}
\description{
This class stores a collection of input data and associated output results from the Structure program.
}
\section{Slots}{

\describe{
\item{\code{summary}}{\code{data.frame}  summary of analyses.}

\item{\code{analyses}}{\code{list} of \code{StructureCollection} objects run using different .}

\item{\code{best}}{\code{numeric} index of \code{StructureCollection} with best negative log-likelihood.}
}}
\seealso{
\code{\link{StructureCollection}}.
}

