% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureCollection.R
\name{StructureCollection}
\alias{StructureCollection}
\title{Create StructureCollection object}
\usage{
StructureCollection(analyses)
}
\arguments{
\item{summary.data}{\code{data.frame}  summary of analyses.}
}
\description{
This function creates a new \code{StructureCollection} object.
}
\details{
The delta-k values are calculated using the method used by Structure Harvestor \url{http://taylor0.biology.ucla.edu/structureHarvester/faq.html.}
}
\seealso{
\code{\link{StructureCollection-class}}, \code{\link{StructureData}}, \code{\link{StructureData}}, \code{\link{StructureResults}}.
}

