% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureData.R
\docType{class}
\name{StructureData-class}
\alias{StructureData-class}
\title{StructureData: An S4 class to represent input data for Structure}
\description{
This class stores input data for the Structure program.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{\code{matrix} with binary loci data. Each row is for a different sample; each column is for a different loci.}

\item{\code{loci.names}}{\code{character} labels used to idenitfy loci.}

\item{\code{sample.names}}{\code{character} labels used to identify labels.}
}}
\seealso{
\code{\link{StructureData}}.
}

