% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureData.R
\name{StructureData}
\alias{StructureData}
\title{Create StructureData object}
\usage{
StructureData(matrix, loci.names, sample.names)
}
\arguments{
\item{matrix}{\code{matrix} with binary loci data. Each row is for a different sample; each column is for a different loci.}

\item{loci.names}{\code{character} labels used to idenitfy loci.}

\item{sample.names}{\code{character} labels used to identify labels.}
}
\description{
This function creates a new \code{StructureData} object.
}
\seealso{
\code{\link{StructureData-class}}.
}

