% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureOpts.R
\docType{class}
\name{StructureOpts-class}
\alias{StructureOpts-class}
\title{StructureOpts: An S4 class to represent Structure parameters}
\description{
This class stores input parameters for the Structure program.
}
\section{Slots}{

\describe{
\item{\code{NUMRUNS}}{\code{numeric} Number of replicate Structure runs. Defaults to 2.}

\item{\code{MAXPOPS}}{\code{numeric} Number of populations assumed. Defaults to 2.}

\item{\code{BURNIN}}{\code{numeric} Length of burnin period. Defaults to 10000.}

\item{\code{NUMREPS}}{\code{numeric} Number of MCMC iterations for inference. Defaults to 20000.}

\item{\code{NOADMIX}}{\code{logical} Do not use admixture model. Defaults to \code{FALSE}.}

\item{\code{ADMBURNIN}}{\code{numeric} Initial period of burnin with admixture model. Defaults to 500.}

\item{\code{SEED}}{\code{numeric} Seed for random number generator. Defaults to NA so a random seed is used.}
}}
\seealso{
\code{\link{StructureOpts}}.
}

