% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureOpts.R
\name{StructureOpts}
\alias{StructureOpts}
\title{Create StructureOpts object}
\usage{
StructureOpts(NUMRUNS = 2, MAXPOPS = 2, BURNIN = 10000, NUMREPS = 20000,
  NOADMIX = FALSE, ADMBURNIN = 500, SEED = sample.int(1e+05, NUMRUNS),
  UPDATEFREQ = max(floor(BURNIN + NUMREPS)/1000, 1))
}
\arguments{
\item{NUMRUNS}{\code{numeric} Number of replicate Structure runs. Defaults to 2.}

\item{MAXPOPS}{\code{numeric} Number of populations assumed. Defaults to 2.}

\item{BURNIN}{\code{numeric} Length of burnin period. Defaults to 10000.}

\item{NUMREPS}{\code{numeric} Number of MCMC iterations for inference. Defaults to 20000.}

\item{NOADMIX}{\code{logical} Do not use admixture model. Defaults to \code{FALSE}.}

\item{ADMBURNIN}{\code{numeric} Initial period of burnin with admixture model. Defaults to 500.}

\item{SEED}{\code{numeric} Seed for random number generator. Defaults to NA so a random seed is used.}

\item{UPDATEFREQ}{\code{numeric} Frequency to store updates to loglikelihood for traceplots. Defaults to yield 1000 frequencies.}
}
\description{
This function creates a new \code{StructureOpts} object.
}
\examples{
StructureOpts(MAXPOPS=2, BURNIN=10000, NUMREPS=20000, NOADMIX=FALSE, ADMBURNIN=500, SEED=NA_real_, UPDATEFREQ=100)
}
\seealso{
\code{\link{StructureOpts-class}}.
}

