% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureReplicate.R
\name{StructureReplicate}
\alias{StructureReplicate}
\title{Create StructureReplicate object}
\usage{
StructureReplicate(loglik, var_loglik, alpha, matrix, sample.names, log)
}
\arguments{
\item{loglik}{\code{numeric} mean negative loglikelihood of model.}

\item{var_loglik}{\code{numeric} variance in negative loglikelihood of model.}

\item{alpha}{\code{numeric} mean value of alpha.}

\item{matrix}{\code{matrix} population membership probabilities. Each row is an individual. Each column is for a different population.}

\item{sample.names}{\code{character} name of samples.}

\item{log}{\code{character} log file.}
}
\value{
\code{\link{StructureReplicate}}.
}
\description{
This function creates a new \code{StructureReplicate} object.
}
\seealso{
\code{\link{StructureReplicate-class}}.
}

