% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureResults.R
\docType{class}
\name{StructureResults-class}
\alias{StructureResults-class}
\title{StructureResults: An S4 class to results from Structure}
\description{
This class stores results from the Structure program.
}
\section{Slots}{

\describe{
\item{\code{summary}}{\code{matrix} object containing average membership probabilities across replicates.}

\item{\code{replicates}}{\code{list} of \code{StructureReplicate} objects.}
}}
\seealso{
\code{\link{StructureResults}}, \code{\link{StructureReplicate}}.
}

