% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureResults.R
\name{StructureResults}
\alias{StructureResults}
\title{Create StructureResults object}
\usage{
StructureResults(replicates, opts = ClumppOpts(), dir = tempdir())
}
\arguments{
\item{replicates}{\code{list} of \code{StructureReplicate} objects.}

\item{opts}{\code{ClumppOpts} object to post-process Structure runs.}

\item{dir}{\code{character} location to run CLUMPP post-processing.}
}
\value{
\code{\link{StructureReplicate}}.
}
\description{
This function creates a new \code{StructureResults} object.
}
\seealso{
\code{\link{StructureReplicate-class}}.
}

