% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot.StructureAnalysis}
\alias{traceplot.StructureCollection}
\title{Traceplot}
\usage{
traceplot(x, parameter, ...)

\method{traceplot}{StructureAnalysis}(x, parameter = "Ln.Like", ...)

\method{traceplot}{StructureCollection}(x, parameter = "Ln.Like",
  k = x@best, ...)
}
\arguments{
\item{x}{\code{StructureCollection} or \code{StructureAnalysis} object.}

\item{parameter}{\code{character} name of parameter to plot. Defaults to 'Ln.Like'. See column names in the \code{mcmc} slot of the \code{StrucreReplicate} objects inside \code{x}.}

\item{...}{not used.}

\item{k}{\code{numeric} indicating number of populations to make traceplot for if \code{x} is a \code{StructureCollection} object.}
}
\description{
This function plots a traceplot showing the negative log-likelihood values in a \code{StructureCollection} object.
}
\seealso{
\code{\link{StructureCollection}}.
}

