% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClumppOpts.R
\name{write.ClumppOpts}
\alias{write.ClumppOpts}
\title{Write Clumpp parameters to file}
\usage{
write.ClumppOpts(x, file)
}
\arguments{
\item{x}{\code{ClumppOpts} object to file.}

\item{file}{\code{character} file path to save file.}
}
\description{
This function writes a \code{ClumppOpts} object to file.
}
\seealso{
\code{ClumppOpts}.
}

