% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurer-internal.R
\name{write.ClumppReplicates}
\alias{write.ClumppReplicates}
\title{Write replicate data to file for clumpp}
\usage{
write.ClumppReplicates(x, file)
}
\arguments{
\item{x}{\code{list} with \code{StructureReplicate} objects.}

\item{file}{\code{character} file path to save data.}
}
\description{
This function writes replicate Structure runs to file for processing in clumpp.
}

