% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureData.R
\name{write.StructureData}
\alias{write.StructureData}
\title{Write data for Structure}
\usage{
write.StructureData(x, file)
}
\arguments{
\item{x}{\code{StructureData} object.}

\item{file}{\code{character} file path to write data.}
}
\description{
This function writes data for Structure.
}
\examples{
x <- read.StructureData(system.file('extdata', 'example_fstat_aflp.dat', package='structurer'))
write.StructureData(x, tempfile(fileext='txt'))
}
\seealso{
\code{StructureData}.
}

