% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureOpts.R
\name{write.StructureOpts}
\alias{write.StructureOpts}
\title{Write Structure parameters to file}
\usage{
write.StructureOpts(x, dir)
}
\arguments{
\item{x}{\code{StructureOpts} object.}

\item{dir}{\code{character} directory to save files.}
}
\description{
This function writes a \code{StructureOpts} object to file.
}
\seealso{
\code{\link{StructureOpts}}.
}

