% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureReplicate.R, R/StructureResults.R, R/StructureAnalysis.R
\name{loglik}
\alias{loglik}
\alias{loglik.StructureAnalysis}
\alias{loglik.StructureReplicate}
\alias{loglik.StructureResults}
\title{Negative Log Likelihood}
\usage{
loglik(x)

\method{loglik}{StructureReplicate}(x)

\method{loglik}{StructureResults}(x)

\method{loglik}{StructureAnalysis}(x)
}
\arguments{
\item{x}{\code{StructureReplicate}, \code{StructureResults}, \code{StructureAnalysis} or \code{StructureCollection} object.}
}
\value{
\code{numeric}.
}
\description{
This function returns the negative loglikelihood of a Structure analysis.
}

