% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureResults.R
\name{StructureResults}
\alias{StructureResults}
\title{Create StructureResults object}
\usage{
StructureResults(summary = NULL, replicates)
}
\arguments{
\item{summary}{\code{data.frame} object containing overall results from Structure replicates.}

\item{replicates}{\code{list} of \code{StructureReplicate} objects.}
}
\value{
\code{\link{StructureReplicate}}.
}
\description{
This function creates a new \code{StructureResults} object.
}
\seealso{
\code{\link{StructureReplicate-class}}.
}

