% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/StructureAnalysis.R, R/StructureCollection.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot.StructureAnalysis}
\alias{traceplot.StructureCollection}
\title{Traceplot}
\usage{
traceplot(x, ...)

\method{traceplot}{StructureAnalysis}(x, ...)

\method{traceplot}{StructureCollection}(x, K = x@best, ...)
}
\arguments{
\item{x}{\code{StructureCollection} or \code{StructureAnalysis} object.}

\item{...}{not used.}
}
\description{
This function plots a traceplot showing the negative log-likelihood values in a \code{StructureCollection} object.
}
\seealso{
\code{\link{StructureCollection}}.
}

